requirements.txt
discord.py
python-dotenv
yt-dlp
pytube
ffmpeg